Add VideoSpeedController and tests
Adds a new controller for VideoSpeed, to help encapsulate items
following SRP.
This also adds a mechanism to track the users most recent speed
selection, so switching from videos / to other media and back will
restore the user's selection on the next video.

+++ app/src/main/java/org/jellyfin/androidtv/ui/playback/PlaybackController.java
@@ -154,7 +154,7 @@ public class PlaybackController {
         mPlaybackMethod = value;
     }
 
-    public void setPlaybackSpeed(float speed){
+    public void setPlaybackSpeed(Double speed){
         mVideoManager.setPlaybackSpeed(speed);
     }
